KRACOEUS-7296 : PR Fixes
