+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/core/DevelopmentProposal.java
@@ -88,8 +88,7 @@ import javax.persistence.*;
 
 import java.sql.Date;
 import java.util.*;
-import org.kuali.rice.kew.api.WorkflowDocument;
-import org.kuali.rice.kew.api.document.DocumentStatus;
+
 @Entity
 @Table(name = "EPS_PROPOSAL")
 @Customizer(DevelopmentProposal.DevelopmentProposalCustomizer.class)
@@ -393,8 +392,7 @@ public class DevelopmentProposal extends KcPersistableBusinessObjectBase impleme
     private transient LegacyNarrativeService narrativeService;
 
     @Transient
-    private transient ProposalPersonBiographyService proposalPersonBiographyService; 
- 
+    private transient ProposalPersonBiographyService proposalPersonBiographyService;
 
     /**
      * Gets the proposalNumberForGG attribute. 
@@ -2024,7 +2022,6 @@ public class DevelopmentProposal extends KcPersistableBusinessObjectBase impleme
      * @return
      */
     public ProposalDevelopmentDocument getProposalDocument() {
-    	  
         if (proposalDocument == null) {
             proposalDocument = new ProposalDevelopmentDocument();
         }
